// src/app/shared/enum/shopping-action-types.enum.ts
export enum ShoppingActionTypes {
  Add = 'Add',
  Add_Success = 'Add Success',
  Add_Failure = 'Add Failure',

  Delete = 'Delete',
  Delete_Success = 'Delete Success',
  Delete_Failure = 'Delete Failure',

  Load_Shopping = 'Load Shopping',
  Load_Success = 'Load Success',
  Load_Failure = 'Load Failure',
}